sync

+++ stm8mcu.h
@@ -10,6 +10,6 @@
 #include "types.h"
 
 // STM8_FREQ defined trhough Makefile
-#define STM8_FREQ_MHZ ((int)(STM8_FREQ / 1000000L))
+#define STM8_FREQ_MHZ ((uint8_t)(STM8_FREQ / 1000000L))
 
 #endif
